services/subscription_service/requirements.txt
anyio==4.4.0
fastapi==0.115.13
pydantic==2.8.2
pydantic-core==2.20.1
starlette==0.41.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.31
httpx==0.27.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.3.0
